feat: reworking tick schedules and using consistent duration objects (#11)
* fix: re-persist city population_cap

The previous cleanup dropped population_cap from the DB on the theory
that the city actor rebuilds it from building registrations on restart.
That's true at the actor layer, but the GetMap read path goes through
store.GetAllCities (the DB), not the actors. With the column gone, the
DB returned PopulationCap=0 for every city. The capital recovered on
the next StreamState push (its actor publishes to the owner), but towns
have no owner and never push, so their cap stayed at 0 in the client
indefinitely.

Restores the column, the queries, and the write path. The actor still
treats itself as the source of truth at runtime; the DB now correctly
reflects that state for read-only consumers.


* feat: per-day rate system, unified 3s tick, Rate proto, pre-built capital farm

Three coupled changes that together fix the "capital pop decreasing on
first boot" UX:

* Rates: production and upkeep are now int64 per-day everywhere. The
  storage unit, the tick math, and the proto all use the same scale,
  removing the per-3s vs per-10s asymmetry in the old code. Balance
  values are chosen as multiples of 28800 (= 86400/3) so per-tick math
  is exact integer division. The four current building rates and pop
  upkeep all preserve today's effective balance.

* Tick cadence: city tick collapses from 10s to 3s so it matches the
  building cadence. Population growth retunes 0.001 -> 0.0003 to keep
  growth feel constant under the shorter tick. Buildings unchanged.

* Wire: introduces a reusable Rate { value, scale } in entity/v1, used
  by City, User, and the new ResourceRate in the ConfigService. Costs
  stay as ResourceAmount (one-shot). Config exposes building_tick_seconds
  and city_tick_seconds for clients that want countdowns.

* Capital UX: player capitals ship with one L1 farm at (StartX+1,
  StartY+1) so the initial 250-pop city is self-sustaining instead of
  starving 5 minutes after boot.


* chore: replace _seconds fields with google.protobuf.Duration

building_tick_seconds, city_tick_seconds, and construction_time switch
from raw int seconds to google.protobuf.Duration. The unit is encoded in
the type instead of the field name, no _seconds suffix anywhere.

Wire format goes from int "3" to Duration "3s" (Connect JSON), which is
self-describing for clients and gets free sub-second precision via the
standard nanos field if ever needed.


* chore: collapse population_cap restore into the original drop migration

00003 was undoing part of 00002 (dropped population_cap, then re-added
it). Since boot wipes and replays the full migration sequence, fold the
restore back into 00002 by simply not dropping the column there in the
first place. Net effect on schema is identical; one fewer migration to
read.


* chore: rename {Building,City}{Production,Backup}Frequency to *TickInterval

CityBackupFrequency was misleading — the city tick runs the food loop,
population growth, and stream push as its primary jobs; the backup
enqueue is just one side effect. BuildingProductionFrequency was
similarly narrow.

Both renamed to {Building,City}TickInterval to honestly describe what
they are (the actor's tick cadence) and to match the proto field names
that already use this language.

Pure rename, no behavior change.


* feat: hide city production intel from non-owners on GetMap and GetCity

Population and starving status stay public (visible from outside the
walls), but the per-day production/upkeep/net rates are economy intel
and should be owner-only.

Adds a HidePrivateCityFields helper and applies it at the two RPC
boundaries where a viewer might see a city they don't own. StreamState
is already owner-only (publishes to *City.Owner) and ListCities only
returns the caller's own cities, so neither needed changes.


* docs: note city field visibility on proto, mapping, and rpc

The City proto now documents which fields are public vs owner-only and
where the restriction is enforced. Matching one-liners at the mapping
helper and the two RPC call sites point back to the same rule so anyone
reading the handler can see why the masking happens.

No code change in this commit, only comments.


---------

+++ internal/domain/city.go
@@ -22,15 +22,15 @@ type City struct {
 	StartY        int      `json:"startY"`
 	Size          int      `json:"size"`
 
-	// FoodProductionRate is food per second this city's own farms produced in
-	// the last tick. FoodUpkeep is food per second this city's population
-	// consumed. NetFoodFlow = production - upkeep (positive = surplus exported
-	// to the user pool; negative = imported from it). Starving is true when
-	// the city could not cover its demand even after drawing from the pool.
-	FoodProductionRate float64 `json:"foodProductionRate"`
-	FoodUpkeep         float64 `json:"foodUpkeep"`
-	NetFoodFlow        float64 `json:"netFoodFlow"`
-	Starving           bool    `json:"starving"`
+	// FoodProductionRate is the food this city's own farms produce per day.
+	// FoodUpkeep is the food this city's population consumes per day. NetFoodFlow
+	// = production - upkeep (positive = surplus exported to the user pool;
+	// negative = imported from it). Starving is true when the city could not
+	// cover its demand even after drawing from the pool.
+	FoodProductionRate int64 `json:"foodProductionRate"`
+	FoodUpkeep         int64 `json:"foodUpkeep"`
+	NetFoodFlow        int64 `json:"netFoodFlow"`
+	Starving           bool  `json:"starving"`
 
 	CreatedAt time.Time `json:"-"`
 	UpdatedAt time.Time `json:"-"`
